Fix issue with integer order_item price and not showing items left after deleteion

+++ db/migrate/20121023023456_create_order_items.rb
@@ -2,7 +2,7 @@ class CreateOrderItems < ActiveRecord::Migration
   def change
     create_table :order_items do |t|
       t.integer :quantity, default: 0, null: false
-      t.integer :price, default: 0.0, null: false
+      t.decimal :price, default: 0.0, null: false
 
       t.integer :order_id, null: false
       t.integer :phone_id, null: false
